feat: add SavedMovies
